requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.7
django-celery-beat==2.6.0
django-ckeditor==6.7.1
django-extensions==3.2.3
django-js-asset==2.2.0
django-jsoneditor==0.2.4
django-jsonform==2.23.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gevent==24.2.1
greenlet==3.0.3
hiredis==3.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
model-bakery==1.19.5
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
pycparser==2.22
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.0.7
referencing==0.35.1
rpds-py==0.20.0
setuptools==71.1.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==6.4.post2
